feat(#861): LCD contrast control moved to control-strip
Extract DIM/LOW/MID/HIGH/MAX preset picker + Shift+[/] keyboard
handler out of the amber display surface into a dedicated
LcdContrastControl component, mounted below the LCD in a new
.lcd-control-strip inside LcdLayout. The amber surface is now
pure readout; contrast lives where it can be clicked comfortably
without reaching into the display.

useLcdContrast store, localStorage persistence, and Shift+[/]
behavior are preserved verbatim from #833.
